Address feedback by @bassosimone
* Test the next words if the word is less than 5 chars in web_connectivity
  test.

* Move the charset detection charset into the correct module and change the
  unittest accordingly.

* Include ooni.common in setup.py

* Various code style fixes
